dc_disney - Add frame and animation range conditions.
Make necessary changes to remove replace ani0001.h (switch animation on frame/animation/control/random conditions with dc_disney. Required removal of velo0001.

+++ data/scripts/dc_disney/condition_health.c
@@ -380,7 +380,7 @@ int dc_disney_check_condition_health_value(void entity)
 */
 int dc_disney_check_health_value(void entity, int threshold)
 {
-	float health_current = get_entity_property(entity, "hp");
+	int health_current = get_entity_property(entity, "hp");
 
 	/* Result exceed value? */
